pyproject.toml
[project]
name = "distill-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.0",
    "autoawq>=0.2.9",
    "datasets>=4.0.0",
    "einops>=0.8.1",
    "torch>=2.8.0",
    "transformers>=4.55.0",
]
